Streamlit/requirements.txt
snowflake-connector-python[pandas]
pandas
numpy
streamlit
requests
plotly
altair
missingno
geopy
prediction
PIL